Clearly missing something on the parser_task.
