+++ commands.sh
@@ -81,6 +81,9 @@ alias tdrs-logs-backend='docker logs tdrs-backend-web-1 -f'
 # Update backend lockfile
 alias tdrs-piplock="tdrs-compose-backend run --rm web pipenv lock"
 
+# A recurring pattern I was doing to get a fresh setup
+alias tdrs-fresh-start='tdrs-stop && docker system prune --volumes && tdrs-start'
+
 # run flake8 against backend source from inside of web container
 tdrs-lint-backend() {
     tdrs-compose-backend run --rm web bash -c "flake8 ."
