+++ tdrs-backend/tdpservice/search_indexes/parsers/preparser.py
@@ -4,6 +4,7 @@ import re
 import logging
 import argparse
 from cerberus import Validator
+from .util import get_record_type
 from . import tanf_parser
 # from .models import ParserLog
 # from tdpservice.data_files.models import DataFile
@@ -11,40 +12,6 @@ from . import tanf_parser
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.DEBUG)
 
-def get_record_type(row):
-    """Get the record type from the row."""
-    if re.match(r'^HEADER.*', row):
-        logger.debug('Matched following row as a header: %s' % row)
-        return 'HE'
-    elif re.match(r'^TRAILER.*', row):
-        logger.debug('Matched following row as a trailer: %s' % row)
-        return 'TR'
-    elif re.match(r'^T1.*', row):
-        logger.debug('Matched following row as data: %s' % row)
-        # if len(row) != 156:
-        #    raise ValueError('T1 row length is not expected length of 156 characters.')
-        return 'T1'
-    elif re.match(r'^T2.*', row):
-        logger.debug('Matched following row as data: %s' % row)
-        return 'T2'
-    elif re.match(r'^T3.*', row):
-        logger.debug('Matched following row as data: %s' % row)
-        return 'T3'
-    elif re.match(r'^T4.*', row):
-        logger.debug('Matched following row as data: %s' % row)
-        return 'T4'
-    elif re.match(r'^T5.*', row):
-        logger.debug('Matched following row as data: %s' % row)
-        return 'T5'
-    elif re.match(r'^T6.*', row):
-        logger.debug('Matched following row as data: %s' % row)
-        return 'T6'
-    elif re.match(r'^T7.*', row):
-        logger.debug('Matched following row as data: %s' % row)
-        return 'T7'
-    else:
-        logger.debug('No match for row: %s' % row)
-        return None
 
 def validate_header(datafile, data_type, given_section):
     """Validate the header line of the datafile."""
@@ -154,6 +121,7 @@ def preparse(datafile, data_type, section):
     """Validate metadata then dispatches file to appropriate parser."""
     # check file type and extension #TODO: this should be done by the frontend but let's verify here
 
+    logger.info("whatwhatwhatwhat")
     # validate header and trailer lines
     header_is_valid, header_errors = validate_header(datafile, data_type, section)
     trailer_is_valid, trailer_errors = validate_trailer(datafile, data_type, section)
