+++ tdrs-backend/tdpservice/search_indexes/parsers/tanf_parser.py
@@ -3,7 +3,7 @@
 import logging
 from ..models import T1  # , T2, T3, T4, T5, T6, T7, ParserLog
 # from django.core.exceptions import ValidationError
-from .preparser import get_record_type
+from .util import get_record_type
 
 logger = logging.getLogger(__name__)
